feat: 修改eslint，prettier弹窗jsx打包位置，解决生产打包构建报错
